backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
celery==5.4.0
redis==5.0.4
yt-dlp==2024.5.27
pytrends==4.9.2
playwright==1.44.0
requests==2.32.2
